hw 04-auth

+++ controllers/contactsControllers.js
@@ -2,8 +2,8 @@ import {Contact} from "../models/contact.js";
 import HttpError from "../helpers/HttpError.js";
 import ctrlWrapper from "../helpers/ctrlWrapper.js";
 
-const getAllContacts = async (_, res) => {
-    const contacts = await Contact.find({});
+const getAllContacts = async (req, res) => {
+    const contacts = await Contact.find({owner: req.user.id});
     res.json(contacts);
 };
 
@@ -11,13 +11,25 @@ const getOneContact = async (req, res) => {
     const {id} = req.params;
     const contact = await Contact.findById(id);
 
-    if (!contact) throw HttpError(404);
+    if (!contact) {
+        throw HttpError(404);
+    }
+
+    if (contact.owner.toString() !== req.user.id) {
+        throw HttpError(404);
+    }
 
     res.json(contact);
 };
 
 const deleteContact = async (req, res) => {
     const {id} = req.params;
+
+    const contact = await Contact.findById(id);
+
+    if (contact.owner.toString() !== req.user.id) {
+        throw HttpError(404);
+    }
 
     const delContact = await Contact.findByIdAndDelete(id);
 
@@ -27,13 +39,27 @@ const deleteContact = async (req, res) => {
 };
 
 const createContact = async (req, res) => {
-    const newContact = await Contact.create(req.body);
+    const contact = {
+        name: req.body.name,
+        email: req.body.email,
+        phone: req.body.phone,
+        favorite: req.body.favorite,
+        owner: req.user.id,
+    };
 
-    res.json(newContact);
+    const result = await Contact.create(contact);
+
+    res.json(result);
 };
 
 const updateContact = async (req, res) => {
     const {id} = req.params;
+
+    const contact = await Contact.findById(id);
+
+    if (contact.owner.toString() !== req.user.id) {
+        throw HttpError(404);
+    }
 
     const updateContact = await Contact.findByIdAndUpdate(id, req.body, {new: true});
